[updates] Move imports from `@expo/config` and `@expo/config-plugins` to `expo`.
