+++ packages/expo-updates/cli/src/utils/modifyConfigAsync.ts
@@ -1,5 +1,5 @@
-import { ExpoConfig, modifyConfigAsync } from '@expo/config';
 import chalk from 'chalk';
+import { ExpoConfig, modifyConfigAsync } from 'expo/config';
 
 import * as Log from './log';
