+++ packages/expo-updates/package.json
@@ -39,8 +39,6 @@
   },
   "dependencies": {
     "@expo/code-signing-certificates": "0.0.5",
-    "@expo/config": "~11.0.10",
-    "@expo/config-plugins": "~10.0.2",
     "@expo/spawn-async": "^1.7.2",
     "arg": "4.1.0",
     "chalk": "^4.1.2",
