+++ packages/expo-updates/utils/src/resolveRuntimeVersionAsync.ts
@@ -1,5 +1,5 @@
-import { getConfig } from '@expo/config';
-import { Updates } from '@expo/config-plugins';
+import { getConfig } from 'expo/config';
+import { Updates } from 'expo/config-plugins';
 import * as Fingerprint from 'expo/fingerprint';
 
 import { createFingerprintAsync } from './createFingerprintAsync';
