+++ packages/expo/config/index.d.ts
