+++ packages/expo/config/index.js
