+++ packages/expo/config/paths.d.ts
@@ -1,0 +1,1 @@
+export * from '@expo/config/paths';
