+++ packages/expo/config/paths.js
@@ -1,0 +1,1 @@
+module.exports = require('@expo/config/paths');
